Don't include plugin test files in the gem file
One of the test files is causing chef omnibus failures on Solaris for some reason, but really there's no reason to ship all these test files in the install artifact. This slims down the gem and install size a good amount. I'm open to a better regex if someone has opinions.

Signed-off-by: Tim Smith <tsmith@chef.io>