Create InSpec Team Working Agreement
This is the written basis of how we "conduct business" within the InSpec team. As our product and project are open source, we should look to publish this for other in the community to see so that they not only know what we are about and how we operate, but so that as they become more involved in the project, they have a shared understanding of how to contribute.